First module
Created a first child module. Modified the pom.xml to mark as parent and
to add the child module.
Created the directory structure for the child module. Created the
pom.xml for the child. The child module has dependencies specific to
that module.
Modified .gitignore to ignore all /target/ directories.