Another module
This time created a module from within IntelliJ using New->Module.This
automatically updated the parent pom with the new module and in the new
module included the parent module information in the pom.xml.
Nothing interesting otherwise.
Empty files created in resources and test/java so the directories would
be included by git, which will not include empty directories.