Add MapStruct Module
Added a mapstruct module. The mapstruct module has dependencies on
jpa-entities and web-api. NOTE: Because web-api is generated code,
IntelliJ will not be able to find it if the target does not exist, so
may need to build first to complete the Mapper code w/o error.
web-app module is updated to depend on the converters (mapstruct)
module. Because the converters module has dependencies on web-api and
jpa-entities, COULD remove them from the web-app module dependencies
list, as it will pick them up from converters. Can also leave them in
for clarity for the next developer who may look at the code.